+++ src/richie/apps/courses/urls.py
@@ -6,7 +6,7 @@ from django.urls import path, re_path
 from rest_framework import routers
 
 from .api import CourseRunsViewSet, course_runs_sync
-from .views import PersonPageAdminAutocomplete
+from .views import PageAdminAutocomplete
 
 ROUTER = routers.SimpleRouter()
 
@@ -16,8 +16,8 @@ ROUTER.register("course-runs", CourseRunsViewSet, "course_runs")
 urlpatterns = ROUTER.urls + [
     re_path("course-runs-sync/?$", course_runs_sync, name="course_run_sync"),
     path(
-        "person-page-admin-autocomplete/",
-        PersonPageAdminAutocomplete.as_view(),
-        name="person-page-admin-autocomplete",
+        "page-admin-autocomplete/<slug:model_name>>/",
+        PageAdminAutocomplete.as_view(),
+        name="page-admin-autocomplete",
     ),
 ]